Implement support for actually unlocking general collectibles via limited case evaluation
`<case>` elements (with `trigger` attributes, instead of being grouped inside `<trigger>` elements) should be added to general_collectibles.xml. They should use conditions (relative to the human player) and can set markers (on the human player) and unlock collectibles.

The cases are treated like hidden cases but only evaluated after all ordinary behaviour updates in situations that use `updateAllBehaviours()` (meaning that a collectible that depends on the hand dealt to the player (before swapping) won't work at the moment).

Priorities can be used to affect the evaluation order, but as I'm counting on the file being manually edited, we could rely entirely on the order of the elements within the file.