Sanako: BlueKoin's sponsorship
This is the biggest update to Sanako since her relaunch, and it should make her a lot more enjoyable to play!

BlueKoin's sponsorship requested:
> It seems Sanako doesn't use her stripping poses in her generic lines. Please use her stripping poses in generic lines.
There was a bit of a misunderstanding here. BlueKoin gave me the benefit of the doubt and presumed Sanako did have stripping poses and they just weren't assigned to the lines he saw. But the tragic truth is that the Sanako rework's biggest shortcoming was that she actually had very few of these, so mostly her clothing was disappearing unassisted!

Two weeks ago, Kuroi finished up posing a beautiful new set of stripping and stripped poses for Sanako, which are included in this commit and assigned to all relevant lines. These new poses are amazing and sexy, and you will love them as much as I do! My sincere thanks go out to Kuroi for this amazing work!

> I'll ask that you please change her stripping lines to have her talk more about her hobbies, interests, and/or anything else we might need to know about who she is.
In addition to her poses, I replaced almost all of Sanako's untargeted stripping lines! This meant removing 39 lines and replacing them with 96 new ones! This relates to some great feedback I got on episode 83 of the "Enter the Inventory" cast:

> "She comes across as not really having much to say and fading into the background a bit if she doesn't have in-depth targets."
Thinking on this, I realized there's no reason why I shouldn't start writing three-part sequences for Sanako's regular lines in the same way I do for targeted ones. Her dialogue shouldn't quite be so plain when she's in the spotlight. Now various aspects of Sanako are explored this way, with each stage having at least two unconditional sequences of lines, with completely new experiences in all stages for if Sanako is on a losing streak, addressing the player by a certain name, or the only woman on the table - with the last of these themes getting a new collectible.

> For a start, please add 15 filter lines wherever you see fit.
I have been working tirelessly on Sanako since BlueKoin's sponsorship, and in that time she's gone from 84 to 143 filtered lines, and I'm not finished yet!