Handle characters with skipped initial layers
Adds logic to handle character costumes that skip stage 0.

Aside from the obvious need to account for skip layers
in character initialization, this also changes certain
dialogue triggers used at the select screen and at
game start to ignore attached stage conditions, so
that they can be used even when initial layers are skipped.

This is a bit of a hack code-wise, but it should be
pretty intuitive as far as writing goes.

Caveat: at least for now, this special-casing does
not extend to `global` cases; they will still need
to be set to the character's true starting stage
after skipping initial layers.