Sanako: Nicole's suggestions, part 3
Adjusted a small number of lines Nicole highlighted in the Sanako-Nagisa interactions that got a little too close to uncomfortable.

Also added a special mother-daughter caution that can appear early in the game instead of a line of dialogue. I've done my best to make playing them together to be as enjoyable as possible for everyone, but it's still something I should give players a heads up about.

This commit puts Sanako above 2000 lines: the second-most I have ever written for a character.