Sanako: Nicole's suggestions, part 4
Wrote a couple of lines to reply to the identity reveals for Okino and Ini. Nothing huge: just acknowledging other people dressing up.

The last problem to solve is how to react to her panties being worn by the player - but only when they're revealed. This is proving more difficult than I expected.