Remaining preset tables credited
Except Wasp (who I overlooked in the previous commit) and our most recent characters, all of these have "TBC" listed because they have not been re-confirmed by their devs. Many will not be possible to confirm, so we can remove that indicator later when the preset table shakeup is complete.