Reduce Linecount Reqs for Character Settings
Summary: I would like to reduce the minimum requirements so that you need 1000 lines in total for your first two settings, and then 500 lines for every setting after that. This means that this still has a higher barrier to initial entry compared to collectibles, but is more forgiving once developers are past the 1500 line mark.
I will set a Final Comment Period for this MR for 2 weeks, ending on Tuesday the 9th of July.

Character Settings have been in the game for 7 months now.
The requirements are currently that you need 1800 lines for the first setting (including 'default' plus your actual setting), and then 1000 lines for every setting after that.

So far, only two characters in the game make use of this feature; Nami, and Yuno.
Nami makes use of it for a toggle between New Game Plus and New Game.
Yuno makes use of it to allow the player to directly enable/disable his dual forfeit with Mari Setogaya.

I would posit that this feature is underused because of the steep requirements.
Personal bias is that, having spread my development across several characters instead of one or two 'taller' characters in terms of linecount, the few cases that I would want to make use of this feature are limited by the steep linecount requirements. For example, enabling/disabling specific modes of play with other characters such as dual forfeits, or specific forfeit variations. If a character has more than one single thing that you want to enable/disable, you would have to pad out their linecount substantially with over 2800 lines just to have two options.

Therefore, in summary, I would like to reduce the minimum requirements so that you need 1000 lines in total for your first two settings, and then 500 lines for every setting after that. This means that this still has a higher barrier to initial entry compared to collectibles, but is more forgiving once developers are past the 1500 line mark. Note that unlike collectibles, you can't use this feature until you have passed the minimum requirement.

To make this into a tangible figure:
Yuno goes from having 2 allowed settings to 4. (Linecount: 2113)
Nami goes from having 9 allowed settings to 18. (Linecount: 9038)
Monika goes from having 3 allowed settings to 7. (Linecount: 3665)
Rose (currently on testing) goes from having 0 allowed settings to 0. (Linecount: 950) She would be unable to get a setting until she breached 1800 lines under current rules.