Sanako: Nicole's suggestions, part 5
Nicole suggested that Sanako should react to the player wearing Sanako's panties collectible, which I was able to implement using the new "Collectible (Visible)" check in the CE. Thank you to Erion for adding it!

Sanako will comment on your own reveal of her panties with a different line based on
- her having just arrived, being fully/mostly dressed, bra visible, panties visible, nude, forfeiting, or finishing
- you being female or small-male, medium-male, or large-male
for a combination of 28 different lines that could play!

Nicole would also like Sanako to react to seeing Nagisa's style of panties worn by the player - and both! - which I'll address in a future (smaller) update. I hope to take any lessons learned with me to write full recognition for Nagisa too.