Ignore trailing space in marker conditions without explicit operators.
I'm pretty sure that the if statement was looking at the wrong capture group and that it was supposed to look at the one containing only the operator, if any, so that trailing space can be ignored (the `\s*` just before `$`).

Also deleted all the unnecessary backslashes. `<=>!@` are not regexp metacharacter, nor do you need backslashes within character classes as long as you put - at the beginning or the end.