Update events.xml to extend summer costumes
Extend the summer event costumes (but not the prefill, highlight or partitioning, or the preset table) for the rest of July. We did this last year with a change to prod-config.xml; this creates a second summer event, coded as `summer_2` that makes use of existing events.xml features.

Summer costumes are still auto-selected by default for this event. We may need to discuss if this is how we want it to behave?


(cherry picked from commit e1d77ea0586eabfe11a517744f10a11e179ec52e)

Co-authored-by: TreeWizard96 <treewizardthegreatest@gmail.com>