Nagisa: More forfeit fantasy lines
If Nagisa is doing her non-play forfeit and the game is over, she needs to draw on a deeper pool of lines. Previously, her chain of fantasy-style lines were being exhausted too quickly, leaving her with nothing to say for about seven phases. This update adds some more lines that will play in this potential circumstance.