New finishing effect (that isn't hidden when hiding the UI)
Put code that correctly displays (or not, as the case may be) this effect in one function (mostly). This simplifies the code and means that we don't have to remember to reset it in loadGameScreen().

Save forfeit properties in rollback states so that the human timer and finishing effect can be correctly shown in rollback, as well as labels (but still not human wardrobe or any hands).

Also fix the bug that the main button says "Cum!" in rollback when the human player was about to.