Major restructuring of UI update code such that updateAllGameVisuals() (which is only called when rolling back) can update the whole UI to reflect any game state
This also involves storing hands and clothing states in RollbackPoint.

Another major change involves keeping gamePhase at the previous phase until advancing the game, in order to be able to correctly display hands revealed or not when exiting rollback. This will also let us include the correct game in bug/feedback reports.

In particular, the new function displayAllHands() and related functions now toggle the loser and tied classes, as well as revealed-cards (shortened from opponent-revealed-cards).

The code is simplified a bit by setting those classes only on the opponent areas instead of on the opponent name labels.

Some code that directly manipulates the UI still remains, especially in the card swap phase.