Nagisa: Added a few lines
Please consider voting for Nagisa in the re-sort. I continue to shower her with love!

Also, this update has the text file for the full interactions between Nagisa, Moon, and the others that can chime in. This is good if you'd rather a quick read than a full SPNatI game. I'll add the transcripts for the interactions with May and Panty in future updates.