Resolve the biggest outstanding issue with unlockable costumes (#158)
Event costumes will now be available during events, but can be unlocked for play outside those events.

To make this work, the alternate_costumes array will be filled with all costume descriptors except those filtered out because of status (offline etc.) and dynamically filtered by .getAvailableCostumes() (renamed from .listUnlockedCostumes() to reflect the fact that it takes not only unlock status into account).

Display of the alternate costume badge also has to be updated dynamically, similar to the epilogue badges. In the next step, we can implement the padlock icon as well. It's probably a good idea to finish #168 first.