Add Glamrock Chica to the main roster as #315!
It's a new day. It's your time to shine. It's time to take your career into your own hands. You built up free time your whole life. Great! Now it's time to put all your eggs into one basket and take a huge gamble on your future. A gamble that comes with a one hundred percent chance of success in some cases. Something almost anyone can do with a limited degree of success. What are we talking about?

We're talking about becoming a SPNATI developer. That's right, character creation and management. Sure, it's a lot of time to invest, but everyone's doing it, and that means it's safe and lucrative.