Add guard to the delay flicker
Problem: In Firefox (and potentially other browsers), the delay to prevent flickering has weird interactions with skipped Card Swap phases, meaning that when a character goes from a Custom Pose or Pose Set back to a String-based image, the String may still have the `display: none` attribute.

Solution: Only hide the <img> if we're STILL on a custom pose at execution time.