Adds initial dynamic language support for the UI
This is a very basic first-steps implementation of dynamic language support for the UI. Although the game's ever-growing dialogue is impossible to translate (at least via traditional means), this is a step towards getting people toward the one language that unites us all: nudity.

Here's what the next steps would be:
- More buttons added to the list of what gets translated: it's literally just "Random Fill" as a test right now.
- More languages supported: it's just Spanish and Japanese for this test.
- Possibly a language modal or settings option: it's based on auto-detect from your browser language right now.

Change your browser language and give it a test!