Correct compensation for the scaleX(0.6) transform of the speech bubble arrows
Draw a right isosceles triangle and make the height _h_ from the hypothenuse the border width (0.2rem). The catheti represent the horizontal and vertical distance across the arrow border (both sqrt(2) · _h_).

Applying an x scale factor of _c_ yields a new right triangle with legs _h_ · sqrt(2) and _ch_ · sqrt(2) and a new height _w_ from the hypothenuse representing the new effective border width.

By using two ways of writing the area of the triangle, we get

_h_ · sqrt(2) · _ch_ · sqrt(2) = _w_ · sqrt((sqrt(2) · h)² + (sqrt(2) · (ch))²)

2_ch_² = _w_ · sqrt(2_h_² + 2_c_²_h_²)

_w_ = 2_ch_² / (sqrt(2) · sqrt(1 + _c_²) · h) =  sqrt(2) · _ch_ / sqrt(1 + _c_²)

Thus sqrt(1 + _c_²) / (_c_ · sqrt(2)) (with _c_ = 0.6) is the factor that we need to multiply the border width 0.2rem by to compensate.