Languages: Request browser's primary language
Previously, we'd request the following data from the browser:
- `navigator.languages`: an array of supported languages
- `navigator.language`: an older variable that is a single variable
- `navigator.userLanguage`: what Internet Explorer used to provide.

However, there was an issue with Google Chrome in that it was supplying the array in an unexpected sequence. For example, if the user had Japanese as their third-ranked language for preference, the output of this function was deciding Japanese was the language of choice, even if options one and two were both English.

By requesting `language` and not `languages`, we can request the browser's primary language. For users that need the UI translated, we can presume that this will be the language they need. This worked well in Chrome. For Firefox, `language` didn't seem to achieve the desired results, so now `languages` is the fallback.

Many other parts of the code are not compatible with Internet Explorer and haven't been since 2018, so we don't need to worry about a special case for IE.