Nagisa: The Underwear Drawer Update
This update allows Nagisa's underwear to be randomized after the first time you've won a game against her. This is done via pose sets that include custom poses in stages 4-6 that have marker-based visibility for alternative images.

Underwear in this update was styled by devs including Fence, Horsecat, Nord, and onboro. My thanks to these champions!

How it works:
- Players that have won a game against Nagisa will have unlocked her "Nagisa's Panties" collectible.
- Before the first round gets underway, the game will check if this collectible is unlocked, and will randomize Nagisa's underwear if so. If not, there is a fallback case that will put her in her default underwear.
- In the case of her waitress costume, the game will instead look for her Waitress Uniform collectible unlock. This means that if you're using the costume but have not unlocked the associated collectible (like when she's featured during the Valentine's event), then you'll get that costume's default underwear.
- For any case that uses stage 4-6, this case gets a Pose Set assigned instead of a regular image. A Pose Set allows the collation of regular poses and Pose Maker poses, choosing the correct one depending on the stage.
- Each custom pose for stages 4-6 uses markers to determine which image to display. (This is tech I first saw used on Lotte's underwear stage where Akko can magic her into lingerie using a marker.)
- For some stretch goals, I also put in the time to make this work for Nagisa's underwear-lowing pose, underwear-lowered poses, and Treasure Scope poses.

There is a built-in debug for forcing a particular underwear type, even if you don't have the collectibles. With Nagisa selected, you can select and deselect specific characters to set the hidden markers to predetermine the result:
- Select and deselect Nayru for the `blue_belle` set (default and waitress)
- Select and deselect Juri for the `mismatched` set (default and waitress)
- Select and deselect Leggy for the rare `white_lace` set (default only)
- Select and deselect Kyu to enforce the original underwear (default and waitress)
Feel free to use and share this information.

The commit has 385 file changes, so I would love any feedback and testing you have after trying it. The most likely error now and in the future is if I assign a traditional pose to a case that includes stages 4-6. As this update has been cooking for a while, it comes with other changes too including new lines, renaming the "shy" pose to "coy" to reflect how it's used, and repositioning of the arms on the "embarrassed_4" pose. I welcome reports on any error, old or new. And positive feedback too if you like what you see!

Credit for inspiration to start working on this update goes to Noko, the Lingerie TCNatI set, and a game called "Unpacking".