Adjust Cyber Lounge
The bg was exported at a higher resolution to preserve some background details, but the tradeoff was that it would start to lag when the game is played in higher resolution, especially on firefox. I have scaled it back down to the side of the next largest bg, making it 3000x1500.