Costume Descriptions
These replace the character's default description when a given costume is selected, as long as the description text box is not empty.

Currently this uses costume attributes in meta.xml. It may be worth changing this so that it uses elements instead: It would then be worth considering moving the costume's name to a <costume-name> element as well. Backwards compatibility would be needed for existing costume names.